Refactored lib/auth*.js files
