+++ static/js/lib/fetch_auth.js
